requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic[email]==2.5.3
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Audio Processing (for integration support)
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3

# Dataset Preparation (Person 1 Extension)
ucimlrepo==0.0.3
pandas==2.1.4
scikit-learn==1.3.2
